move Levenshtein calculations to JournalUtils from OdinsHamr
