+++ dspace/modules/api/src/main/java/org/dspace/curate/OdinsHamr.java
@@ -40,6 +40,7 @@ import org.dspace.authority.orcid.model.Bio;
 import org.dspace.authority.orcid.model.BioName;
 import org.dspace.authority.AuthorityValueGenerator;
 import org.dspace.content.authority.Choices;
+import org.dspace.JournalUtils;
 
 import org.apache.log4j.Logger;
 
@@ -231,35 +232,7 @@ public class OdinsHamr extends AbstractCurationTask {
         return Curator.CURATE_SUCCESS;
 	}
 
-    /**
-       Computes a minimum between three integers.
-    **/
-    private static int minimum(int a, int b, int c) {
-	    return Math.min(Math.min(a, b), c);
-    }
 
-    /**
-       Levenshtein distance algorithm, borrowed from
-       http://en.wikibooks.org/wiki/Algorithm_Implementation/Strings/Levenshtein_distance
-    **/
-    public static int computeLevenshteinDistance(CharSequence str1, CharSequence str2) {
-        int[][] distance = new int[str1.length() + 1][str2.length() + 1];
-
-        for (int i = 0; i <= str1.length(); i++)
-            distance[i][0] = i;
-        for (int j = 1; j <= str2.length(); j++)
-            distance[0][j] = j;
-
-        for (int i = 1; i <= str1.length(); i++)
-            for (int j = 1; j <= str2.length(); j++)
-            distance[i][j] = minimum(
-                         distance[i - 1][j] + 1,
-                         distance[i][j - 1] + 1,
-                         distance[i - 1][j - 1] + ((str1.charAt(i - 1) == str2.charAt(j - 1)) ? 0 : 1));
-
-        return distance[str1.length()][str2.length()];
-    }
-    
     /**
        Scores the correspondence between two author names.
     **/
@@ -270,7 +243,7 @@ public class OdinsHamr extends AbstractCurationTask {
         }
 
         int maxlen = Math.max(getName(bio1).length(), getName(bio2).length());
-        int editlen = computeLevenshteinDistance(getName(bio1), getName(bio2));
+        int editlen = JournalUtils.computeLevenshteinDistance(getName(bio1), getName(bio2));
 
         return (double)(maxlen-editlen)/(double)maxlen;
     }
